+++ src/controllers/api/utility.js
@@ -1,6 +1,3 @@
-const redis = require('redis');
-const { redisUrl } = require('../../util');
-
 const ClassAPIResource = data => {
   const mapped = data.map(item => {
     return {
@@ -35,15 +32,8 @@ const escapeRegExp = string => {
   return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'); // $& means the whole matched string
 };
 
-const redisClient = redis.createClient(redisUrl);
-const closeRedisClient = callback => {
-  redisClient.quit(callback);
-};
-
 module.exports = {
   NamedAPIResource,
   ClassAPIResource,
-  escapeRegExp,
-  redisClient,
-  closeRedisClient
+  escapeRegExp
 };
