+++ src/tests/integration/server.itest.js
@@ -1,8 +1,7 @@
 const mongoose = require('mongoose');
 const request = require('supertest');
 const app = require('../../server');
-const { mongodbUri } = require('../../util');
-const { closeRedisClient } = require('../../controllers/api/utility');
+const { mongodbUri, redisClient, closeRedisClient } = require('../../util');
 
 beforeAll(async () => {
   await mongoose.connect(mongodbUri, { useNewUrlParser: true, useUnifiedTopology: true });
